requirements.txt
streamlit
google-genai
chromadb
python-dotenv
tqdm
datasets
# local embedding için
sentence-transformers
torch
